Feat : 주문 완료 페이지 상품 품절 시 장바구니,상품리스트,상품 상세페이지 변경 추가
Fix : 상품 등록 대표 이미지 체크 수정
장바구니 상품 삭제 수정

DB MySQL로 변경

+++ src/main/java/com/developer/foodfit/service/OrderItemService.java
@@ -32,13 +32,15 @@ public class OrderItemService {
         orderItemRepository.save(orderItem);
 
         //상품 수량 수정
-        itemService.updateStockNumber(orderItem.getItem().getId(),request.getCount());
+        itemService.updateItemCount(item.getId(),request.getCount());
 
         //장바구니 상품 삭제
         Item itemId = itemService.findById(item.getId());
         Cart Cart = cartService.findCartId(order.getUser().getId());
-        CartItem cartItemId = cartItemService.findCartItemId(Cart.getId(), itemId.getId());
-        cartItemService.delete(cartItemId.getId());
+        if(Cart!=null && request.getOrderType().equals("cart")){
+            CartItem cartItemId = cartItemService.findCartItemId(Cart.getId(), itemId.getId());
+            cartItemService.delete(cartItemId.getId());
+        }
 
         return orderItem;
     }
